Use the token vs. string distinction for the null policy
The structured headers specification provides a distinction between unquoted tokens, and quoted strings. We can leverage that distinction by saying that no policy is indicated via a special token, whereas policy IDs are indicated via strings. This avoids the awkward reserving of a single string as special. For example we no longer need to treat "id": "none" in the JSON as a parse error, because the string "none" is no longer overlapping with the token space.

Additionally, this updates the token to be null, instead of none, since it is natural to talk about the null origin policy.

Note that all the examples previously did not quote policy names, so it appears we were unintentionally using tokens for them. This was simply because I was unaware that strings were supposed to be quoted, and that the string/token distinction existed. It makes sense for the policy names to be strings, not tokens, and this commit updates the examples accordingly.