Change the spec representation of the null policy
Previously it was represented as an origin policy whose IDs were the list containing the single item null. This was done because it was somewhat convenient for writing the algorithm in previous revisions of "fetch an origin's origin policy", i.e. we could compare allowedIds (which might contain a null from the header) with cachedPolicy's IDs (which might contain null if the cached policy was the null policy).

However, this was overly clever and confusing. It caused any algorithms involving ID comparison to need careful inspection to make sure they dealt with such an exceptional case. And, it gets in the way when you want to do things like reflect the origin policy IDs to JavaScript, per #48.

So instead, this commit moves to a model where the null origin policy has no IDs. With the current revision of "fetch an origin's origin policy", this seems to purely simplify things.